+++ src/scave/sqlitevectorfileexporter.cc
@@ -145,9 +145,9 @@ void SqliteVectorFileExporter::saveResults(const std::string& fileName, ResultFi
             const VectorResult *vector = manager->getVector(id);
             void *vectorHandle = vectorHandles[i];
             XYArray *array = xyArrays[i];
-            int length = array->length();
+            size_t length = array->length();
             bool hasPreciseX = array->hasPreciseX();
-            for (int j = 0; j < length; j++) {
+            for (size_t j = 0; j < length; j++) {
                 const BigDecimal time = hasPreciseX ? array->getPreciseX(j) : BigDecimal(array->getX(j));
                 if (!time.isSpecial())
                     writer.recordInVector(vectorHandle, array->getEventNumber(j), time.getMantissaForScale(simtimeScaleExp), array->getY(j));
